Move menu instantiation to gui-scene.ts
ItemMenu is now instantiated in create() of gui-scene.ts. Create
guiPack.json to preload all necessary images in gui-scene.ts

Now menu is affected by the existing pause logic. That means the
budget won't be increased during the game (main-scene.ts) is paused,
but the player is still able to buy items.