Change distributeNewRoles method
The method now checks wether the agent is already a health worker
or a police officer. If not, it will overwrite the specific agent
with a new instance of the specific agent subclass Police or
HealthWorker. Health workers can only have the state CURE but
police officers can be infected. This is why, if they where infected
when being promoted to be a police officer they keep their state.