Adds different difficulty levels to the game
There is one json-file per difficulty level which contains the
appropriate values for the game variables of `stats.ts`. The core
logic of different difficulty levels is accessed in
`start-menu-scene.ts` and implemented in `stats.ts`.
For that a new enum (`difficultyLevell.ts` in util) was created
and the constructor as well as `getInstance()` were customized.

Additional activities:
Added missing documentation to `healthStates.ts`