Fix some bugs and 'redocumented' `button-container.ts`
Issue #106:
Destroy the priceText of research in `button-container.ts#buttonAnimations`
if the research level reached ten

Issue #108, #112:
These issues were caused due to the same bug in
`button-container.ts#buttonAnimations`. There were made some logical
mistakes in the if-else-statement. For this reason, e.g. the lockdown
button invoked buyHealthworkers without price and amount parameters.
That led to that the nbrHW in `stats.ts` was set to undefined. Since
this number is used for the salary calculation of health workers it has
also become undefined (as like as the budget).